# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is all the APIs for the application ADA-Restaurant-Booking-System",
    "title": "ADA-Restaurant-Booking-System",
    "version": "v1"
  },
  "host": "localhost:3000",
  "schemes": [
    "http"
  ],
  "basePath": "/",
  "paths": {
    "/tables/get-bookings-of-section": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "Current section bookings",
        "description": "This method is called in the home page from front-end. It is also called when user find a section on a specific date. Returns a list of booking for a specific section on one date",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "section",
            "in": "query",
            "description": "1/2/3 means Morning/Lunch/Dinner section",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "bookingDate",
            "in": "query",
            "description": "the bookind date",
            "required": true,
            "type": "string",
            "format": "DD/MM/YYYY"
          },
          {
            "name": "token",
            "in": "query",
            "description": "An unauthorised and authorised users could use this function; however, when an authorised use the token, it will return the selected table from this user",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Tables is returned successfully",
            "schema": {
              "$ref": "#/definitions/bookingsForSection"
            }
          },
          "404": {
            "description": "User is not found"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/tables/auth/get-current-unconfirmed-bookings": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "Get unconfirmed booking for this user",
        "description": "This method is called when user in page \"confirm booking\" to return a list of table(s) is/are reserved by this current user. To access this API, user has to be authorised.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "section",
            "in": "query",
            "description": "1/2/3 means Morning/Lunch/Dinner section",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "bookingDate",
            "in": "query",
            "description": "the bookind date",
            "required": true,
            "type": "string",
            "format": "DD/MM/YYYY"
          },
          {
            "name": "token",
            "in": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Tables is returned successfully",
            "schema": {
              "$ref": "#/definitions/reservedBooking"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/tables/auth/users-booking-history-list": {
      "get": {
        "tags": [
          "Tables"
        ],
        "summary": "Get booking history",
        "description": "Get history booking and paging for booking history. It will group the history by sections in a specific day.",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "p",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "token",
            "in": "query",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Get booking history successfully",
            "schema": {
              "$ref": "#/definitions/historyBooking"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/tables/auth/reserve-table": {
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Reserve table",
        "description": "This method is called mullitple times in front-end if user booking many table.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/inputObjForReserveBooking"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Table is reserved successfully",
            "schema": {
              "$ref": "#/definitions/info"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/tables/auth/confirm-reserved-bookings": {
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Confirm reserved bookings",
        "description": "This method get an array of reserved booking list. The email provided from the used will be recived a email about the booking detail.",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "reservedBookings",
            "in": "body",
            "required": true,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/userInfoForBooking"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Booking is confirmed successfully",
            "schema": {
              "$ref": "#/definitions/info"
            }
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/tables/auth/delete-reserved-booking": {
      "post": {
        "tags": [
          "Tables"
        ],
        "summary": "Delete reserved booking",
        "description": "This api is called when 1. the user leaves the \"booking confirmed\" screen under 10 minutes. 2. when the user reserves a list of table, when a table is failed this api will be called to removed the booking",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "type": "string",
            "required": true
          },
          {
            "name": "bookingId",
            "in": "body",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Reserved table is delete successfully"
          },
          "401": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/users/sign-up": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register new account",
        "description": "Register new account",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "firstName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "lastName",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "phoneNumber",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User signs up successfully"
          },
          "406": {
            "description": "The email is registeed"
          },
          "409": {
            "description": "User not authenticated"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    },
    "/users/sign-in": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Sign-in",
        "description": "Sign in to confirm booking and view booking history",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "formData",
            "required": true,
            "type": "string"
          },
          {
            "name": "password",
            "in": "formData",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "User login successfully"
          },
          "400": {
            "description": "Invalid login credentials"
          },
          "500": {
            "description": "An error occurred!"
          }
        }
      }
    }
  },
  "definitions": {
    "table": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "capacity": {
          "type": "integer",
          "format": "int32"
        },
        "isSmoking": {
          "type": "boolean"
        },
        "location": {
          "type": "string",
          "enum": [
            "top-right",
            "top-left",
            "right-bar",
            "left-kitchen",
            "right-kitchen"
          ]
        }
      }
    },
    "booking": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "user": {
          "type": "string",
          "description": "the id of user"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "tableId": {
          "type": "object",
          "$ref": "#/definitions/table"
        }
      }
    },
    "bookingDetail": {
      "type": "object",
      "properties": {
        "bookingDate": {
          "type": "string",
          "format": "date-time"
        },
        "confirmDeadline": {
          "type": "string",
          "format": "date-time"
        },
        "createDate": {
          "type": "string",
          "format": "date-time"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "requirement": {
          "type": "string"
        },
        "section": {
          "type": "integer"
        },
        "status": {
          "type": "integer"
        },
        "tableId": {
          "type": "object",
          "$ref": "#/definitions/table"
        },
        "_id": {
          "type": "string"
        }
      }
    },
    "userInfoForBooking": {
      "type": "object",
      "properties": {
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "requirement": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        }
      }
    },
    "bookingsForSection": {
      "type": "object",
      "properties": {
        "reservedTables": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/booking"
          }
        },
        "reservedTablesByUserId": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/booking"
          }
        },
        "tableList": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/booking"
          }
        }
      }
    },
    "reservedBooking": {
      "type": "object",
      "properties": {
        "reservedTables": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/booking"
          }
        },
        "user": {
          "type": "object",
          "$ref": "#/definitions/user"
        }
      }
    },
    "historyBooking": {
      "type": "object",
      "properties": {
        "bookings": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/bookingDetail"
          }
        },
        "page": {
          "type": "string"
        },
        "total": {
          "type": "integer"
        }
      }
    },
    "inputObjForReserveBooking": {
      "type": "object",
      "properties": {
        "section": {
          "description": "1/2/3 means Morning/Lunch/Dinner section",
          "type": "integer"
        },
        "bookingDate": {
          "description": "the booking date",
          "type": "string",
          "format": "DD/MM/YYYY"
        },
        "bookingTable": {
          "type": "object",
          "$ref": "#/definitions/table"
        }
      }
    },
    "user": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "_id": {
          "type": "string"
        }
      }
    },
    "info": {
      "type": "object",
      "properties": {
        "sucess": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "error": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "error": {
          "type": "string"
        }
      }
    }
  }
}